+++ bin/react-create.js
@@ -1,0 +1,8 @@
+#!/usr/bin/env node
+
+const program = require('commander');
+
+require('../lib/create')({
+  cwd: process.cwd(),
+  defaultConfigDir: process.env.USERPROFILE
+});